test(jsx2mp): add props.identifier test case
